bigdata/requirements.txt
requests==2.24.0
flask==1.1.2
<<<<<<< HEAD
flask-pymongo==2.3.0
gunicorn==20.0.4
=======
flask-pymongo==2.3.0
>>>>>>> gitlab/feature/keyword
